+++ crates/proof-of-sql/src/sql/evm_proof_plan/plans/evm_aggregate_exec.rs
@@ -12,6 +12,7 @@ use crate::{
         proof_plans::AggregateExec,
     },
 };
+use alloc::{boxed::Box, string::String, vec::Vec};
 use serde::{Deserialize, Serialize};
 use sqlparser::ast::Ident;
