+++ crates/proof-of-sql/src/sql/evm_proof_plan/plans/evm_group_by_exec.rs
@@ -12,6 +12,7 @@ use crate::{
         proof_plans::GroupByExec,
     },
 };
+use alloc::{string::String, vec::Vec};
 use serde::{Deserialize, Serialize};
 use sqlparser::ast::Ident;
