+++ crates/proof-of-sql/src/sql/evm_proof_plan/plans/evm_sort_merge_join_exec.rs
@@ -8,6 +8,7 @@ use crate::{
         proof_plans::SortMergeJoinExec,
     },
 };
+use alloc::{boxed::Box, string::String, vec::Vec};
 use serde::{Deserialize, Serialize};
 use sqlparser::ast::Ident;
