+++ crates/proof-of-sql/src/sql/evm_proof_plan/plans/evm_table_exec.rs
@@ -8,6 +8,7 @@ use crate::{
         proof_plans::TableExec,
     },
 };
+use alloc::vec::Vec;
 use serde::{Deserialize, Serialize};
 
 /// Represents a table execution plan in EVM.
